+++ acropolis-persistence/src/main/kotlin/org/ephyra/acropolis/persistence/api/persistence/NetworkPersistence.kt
@@ -20,7 +20,7 @@ class NetworkPersistence {
     }
 
     fun findByName(name: String, projectId: Long): NetworkEntity? {
-        val entity = repo.findByNameAndProjectId(name, projectId.toInt())
+        val entity = repo.findByNameAndProjectId(name, projectId)
         return if (entity.isPresent) entity.get() else null
     }
