+++ acropolis-persistence/src/main/kotlin/org/ephyra/acropolis/persistence/impl/NetworkRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository
 import java.util.*
 
 interface NetworkRepository : CrudRepository<NetworkEntity, Long> {
-    fun findByNameAndProjectId(name: String, projectId: Int): Optional<NetworkEntity>
+    fun findByNameAndProjectId(name: String, projectId: Long): Optional<NetworkEntity>
 }
